backend/requirements.txt
kombu==5.2.4
psycopg2==2.9.3
pytz==2022.1
redis==4.2.0
requests==2.27.1
SQLAlchemy==1.4.32
pyexecjs==1.5.1
bs4==0.0.1
pycrypto==2.6.1
bcrypt==3.2.0
Werkzeug==0.16.0
lxml==4.9.1
